complete

+++ shoppingcart/src/main/java/com/lambdaschool/shoppingcart/controllers/CartController.java
@@ -1,19 +1,18 @@
 package com.lambdaschool.shoppingcart.controllers;
 
+import com.lambdaschool.shoppingcart.exceptions.ResourceNotFoundException;
 import com.lambdaschool.shoppingcart.models.Cart;
 import com.lambdaschool.shoppingcart.models.Product;
 import com.lambdaschool.shoppingcart.models.User;
+import com.lambdaschool.shoppingcart.repositories.UserRepository;
 import com.lambdaschool.shoppingcart.services.CartService;
+import com.lambdaschool.shoppingcart.services.HelperFunctions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
@@ -24,9 +23,20 @@ public class CartController
     @Autowired
     private CartService cartService;
 
+    @Autowired
+    private UserRepository userrepos;
+
+    @Autowired
+    private HelperFunctions helperFunctions;
+
     @GetMapping(value = "/user", produces = {"application/json"})
-    public ResponseEntity<?> listAllCarts(@PathVariable long userid)
-    {
+    public ResponseEntity<?> listAllCarts() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        User authenticatedUser = userrepos.findByUsername(authentication.getName().toLowerCase());
+
+        Long userid = authenticatedUser.getUserid();
+
         List<Cart> myCarts = cartService.findAllByUserId(userid);
         return new ResponseEntity<>(myCarts, HttpStatus.OK);
     }
@@ -39,7 +49,7 @@ public class CartController
     {
         Cart p = cartService.findCartById(cartId);
         return new ResponseEntity<>(p,
-                                    HttpStatus.OK);
+                HttpStatus.OK);
     }
 
     @PostMapping(value = "/create/user/{userid}/product/{productid}")
@@ -58,29 +68,50 @@ public class CartController
 
     @PutMapping(value = "/update/cart/{cartid}/product/{productid}")
     public ResponseEntity<?> updateCart(@PathVariable long cartid,
-                                        @PathVariable long productid)
-    {
-        Cart dataCart = new Cart();
-        dataCart.setCartid(cartid);
+                                        @PathVariable long productid) {
 
-        Product dataProduct = new Product();
-        dataProduct.setProductid(productid);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
-        cartService.save(dataCart, dataProduct);
-        return new ResponseEntity<>(HttpStatus.OK);
+        User authenticatedUser = userrepos.findByUsername(authentication.getName().toLowerCase());
+
+        if (helperFunctions.isAuthorizedToMakeChange(authenticatedUser.getUsername())) {
+
+            Cart dataCart = new Cart();
+            dataCart.setCartid(cartid);
+
+            Product dataProduct = new Product();
+            dataProduct.setProductid(productid);
+
+            cartService.save(dataCart, dataProduct);
+            return new ResponseEntity<>(HttpStatus.OK);
+
+        } else {
+
+            throw new ResourceNotFoundException("This user is not authorized to make this change.");
+        }
     }
 
     @DeleteMapping(value = "/delete/cart/{cartid}/product/{productid}")
     public ResponseEntity<?> deleteFromCart(@PathVariable long cartid,
-                                            @PathVariable long productid)
-    {
-        Cart dataCart = new Cart();
-        dataCart.setCartid(cartid);
+                                            @PathVariable long productid) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
-        Product dataProduct = new Product();
-        dataProduct.setProductid(productid);
+        User authenticatedUser = userrepos.findByUsername(authentication.getName().toLowerCase());
 
-        cartService.delete(dataCart, dataProduct);
-        return new ResponseEntity<>(HttpStatus.OK);
+        if (helperFunctions.isAuthorizedToMakeChange(authenticatedUser.getUsername())) {
+
+            Cart dataCart = new Cart();
+            dataCart.setCartid(cartid);
+
+            Product dataProduct = new Product();
+            dataProduct.setProductid(productid);
+
+            cartService.delete(dataCart, dataProduct);
+            return new ResponseEntity<>(HttpStatus.OK);
+
+        } else {
+
+            throw new ResourceNotFoundException("This user is not authorized to make this change.");
+        }
     }
 }
